+++ src/components/Tooltip/index.tsx
@@ -1,46 +1,46 @@
 'use client';
 
-import { useMemo } from 'react';
+import { createPortal } from 'react-dom';
 
-import { formatTimeToString } from '@helpers/time';
 import { useSelector } from '@xstate/store/react';
 import { store } from './store';
 
-// TODO this should be a portal
 export const Tooltip = () => {
-  const time = useSelector(store, (state) => state.context.time);
+  const text = useSelector(store, (state) => state.context.text);
   const x = useSelector(store, (state) => state.context.pos[0]);
   const y = useSelector(store, (state) => state.context.pos[1]);
+  const isVisible = useSelector(store, (state) => state.context.isVisible);
 
-  const timeString = useMemo(() => formatTimeToString(time), [time]);
+  if (!isVisible) return null;
 
-  if (x === -1) return null;
-
-  return (
+  return createPortal(
     <div
       className={`
         absolute 
         z-[900] 
-        w-[6.8rem] 
+        min-w-[6.8rem] 
+        max-w-[10rem]
         bg-tooltip text-black 
         font-mono text-sm text-center 
-        flex items-center justify-center`}
+        flex items-center justify-center
+        pointer-events-none`}
       style={{
         // safari fix - https://stackoverflow.com/a/62934196/2377677
         display: 'inline-block',
         top: y,
         left: x,
-        transform: 'translateX(-35%)',
-        borderRadius: 10
+        borderRadius: 10,
+        transform: 'translateX(-50%) translateY(-50%)'
       }}
     >
-      {timeString}
+      {text}
       <div
         className={`tooltip-arrow absolute left-1/2 top-full -translate-x-1/2 -mt-0 
                           border-solid border-t-8 border-x-8 border-b-0
                           border-t-tooltip border-x-transparent`}
         aria-hidden='true'
       />
-    </div>
+    </div>,
+    document.body
   );
 };